@@ -22,9 +22,16 @@ module JobOfferStateTimestamp
   end
 
   def rebuild_timestamp!(state_name)
+    enum_val = JobOffer.states[state_name]
     target_audit = audits.reorder(version: :desc).detect do |audit|
-      state_has_changed = audit.audited_changes.keys.include?('state')
-      state_has_changed && audit.audited_changes['state'].last == state_name
+      state_val = audit.audited_changes['state']
+      return false if state_val.nil?
+
+      if state_val.is_a?(Array)
+        state_val.last == enum_val
+      elsif state_val.is_a?(String)
+        state_val == enum_val
+      end
     end
     update_column "#{state_name}_at", target_audit.created_at if target_audit
   end
