@@ -10,9 +10,11 @@ require 'active_record/railtie'
 require 'active_storage/engine'
 require 'action_controller/railtie'
 require 'action_mailer/railtie'
+require 'action_mailbox/engine'
+require 'action_text/engine'
 require 'action_view/railtie'
 # require 'action_cable/engine'
-# require "sprockets/railtie"
+# require 'sprockets/railtie'
 require 'rails/test_unit/railtie'
 
 # Require the gems listed in Gemfile, including any gems
@@ -22,7 +24,7 @@ Bundler.require(*Rails.groups)
 module Civilsdeladefense
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 5.2
+    config.load_defaults 6.0
 
     config.time_zone = 'Paris'
 
