@@ -5,9 +5,9 @@ FactoryBot.define do
     first_name { Faker::Name.first_name }
     last_name { Faker::Name.last_name }
     email { Faker::Internet.safe_email }
-    role 'bant'
-    very_first_account true
-    password 'f4k3p455w0rD!'
+    role { 'bant' }
+    very_first_account { true }
+    password { 'f4k3p455w0rD!' }
     confirmed_at { DateTime.now }
   end
 end
