@@ -2,8 +2,8 @@
 
 FactoryBot.define do
   factory :email_template do
-    title 'MyString'
-    subject 'MyString'
-    body 'MyText'
+    title { 'MyString' }
+    subject { 'MyString' }
+    body { 'MyText' }
   end
 end
