@@ -2,8 +2,8 @@
 
 FactoryBot.define do
   factory :email, class: 'Email' do
-    subject 'New email'
-    body 'MyText'
+    subject { 'New email' }
+    body { 'MyText' }
     job_application
     association :sender, factory: :administrator
   end
