@@ -4,8 +4,8 @@ FactoryBot.define do
   factory :job_application do
     job_offer
     user
-    terms_of_service true
-    certify_majority true
+    terms_of_service { true }
+    certify_majority { true }
 
     before(:create) do |job_application|
       job_application.personal_profile = create(:personal_profile,
