@@ -2,8 +2,8 @@
 
 FactoryBot.define do
   factory :job_offer_actor do
-    job_offer nil
-    administrator nil
-    role ''
+    job_offer { nil }
+    administrator { nil }
+    role { '' }
   end
 end
