@@ -2,7 +2,7 @@
 
 FactoryBot.define do
   factory :message, class: 'Message' do
-    body 'MyText'
+    body { 'MyText' }
     job_application
     administrator
   end
