@@ -2,9 +2,9 @@
 
 FactoryBot.define do
   factory :salary_range do
-    estimate_annual_salary_gross '42 €'
-    professional_category nil
-    experience_level nil
-    sector nil
+    estimate_annual_salary_gross { '42 €' }
+    professional_category { nil }
+    experience_level { nil }
+    sector { nil }
   end
 end
