@@ -2,8 +2,8 @@
 
 FactoryBot.define do
   factory :user_file do
-    content 'MyString'
-    user nil
-    job_application_file_type nil
+    content { 'MyString' }
+    user { nil }
+    job_application_file_type { nil }
   end
 end
