@@ -5,7 +5,7 @@ FactoryBot.define do
     first_name { Faker::Name.first_name }
     last_name { Faker::Name.last_name }
     email { Faker::Internet.safe_email }
-    password 'f4k3p455w0rD!'
+    password { 'f4k3p455w0rD!' }
     after(:create) do |user|
       user.personal_profile = create(:personal_profile, personal_profileable: user)
     end
